fixed pattern for folder names
