+++ src/main/java/com/bookStore/BookStore/config/TestConfig.java
@@ -1,22 +1,22 @@
-package com.bookStore.BookStore.Config;
+package com.bookStore.BookStore.config;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 
-import com.bookStore.BookStore.Service.DBService;
+import com.bookStore.BookStore.services.DBService;
 
 @Configuration
-@Profile("prod")
-public class DevConfig {
+@Profile("test")
+public class TestConfig {
 
     @Autowired
-    private DBService service;
+    private DBService dbService;
 
     @Bean
     public void instanciaDB() {
-        this.service.instanciaDados();
+        this.dbService.instanciaDados();
     }
 
 }
