+++ src/main/java/com/bookStore/BookStore/services/BookService.java
@@ -1,4 +1,4 @@
-package com.bookStore.BookStore.Service;
+package com.bookStore.BookStore.services;
 
 import java.util.List;
 import java.util.Optional;
@@ -6,12 +6,12 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.bookStore.BookStore.Repository.BookRepository;
-import com.bookStore.BookStore.Service.exceptions.DateIntegrityViolention;
-import com.bookStore.BookStore.Service.exceptions.ObjectNotFound;
 import com.bookStore.BookStore.dto.BookDto;
 import com.bookStore.BookStore.entities.Book;
 import com.bookStore.BookStore.entities.Category;
+import com.bookStore.BookStore.repositories.BookRepository;
+import com.bookStore.BookStore.services.exceptions.DateIntegrityViolention;
+import com.bookStore.BookStore.services.exceptions.ObjectNotFound;
 
 @Service
 public class BookService {
