+++ src/main/java/com/bookStore/BookStore/services/CategoryService.java
@@ -1,4 +1,4 @@
-package com.bookStore.BookStore.Service;
+package com.bookStore.BookStore.services;
 
 import java.util.List;
 import java.util.Optional;
@@ -6,11 +6,11 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.bookStore.BookStore.Repository.CategoryRepository;
-import com.bookStore.BookStore.Service.exceptions.DateIntegrityViolention;
-import com.bookStore.BookStore.Service.exceptions.ObjectNotFound;
 import com.bookStore.BookStore.dto.CategoryDto;
 import com.bookStore.BookStore.entities.Category;
+import com.bookStore.BookStore.repositories.CategoryRepository;
+import com.bookStore.BookStore.services.exceptions.DateIntegrityViolention;
+import com.bookStore.BookStore.services.exceptions.ObjectNotFound;
 
 @Service
 public class CategoryService {
